requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1